ai_ml/requirements.txt
# =============================================================================
# HackUTA 2025 - Consolidated Python Dependencies
# =============================================================================
# This file contains all Python dependencies for the entire project.
# It consolidates requirements from backend/ and ai_ml/ components.

# =============================================================================
# BACKEND COMPONENT DEPENDENCIES
# =============================================================================

# Core FastAPI dependencies
fastapi==0.118.0
uvicorn==0.37.0
starlette==0.48.0

# Data validation and serialization
pydantic==2.11.10
pydantic-core==2.33.2

# Google AI/ML services
google-genai==1.41.0
google-auth==2.41.1

# HTTP client and networking
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
urllib3==2.5.0

# Async support
anyio==4.11.0
sniffio==1.3.1
websockets==15.0.1

# Authentication and security
rsa==4.9.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
cachetools==6.2.0

# Utilities
click==8.3.0
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2

# Text processing
charset-normalizer==3.4.3
idna==3.10
certifi==2025.8.3

# Windows-specific (optional)
colorama==0.4.6; sys_platform == 'win32'

# Type annotations
annotated-types==0.7.0

# Low-level HTTP
h11==0.16.0

# =============================================================================
# AI/ML COMPONENT DEPENDENCIES
# =============================================================================

# PDF processing and document analysis
pdfplumber
pdf2image
pytesseract
Pillow

# Google Generative AI (alternative to google-genai)
google-generativeai

# FastAPI file upload support
python-multipart

# Environment variables
python-dotenv

# Voice synthesis
elevenlabs

# =============================================================================
# SHARED DEPENDENCIES
# =============================================================================
# Note: Some dependencies may be shared between components
# Version conflicts should be resolved by using the latest compatible version
